fixup! feat(suite-native): dashboard staking banner implementation
