+++ suite-native/intl/src/messages.ts
@@ -155,9 +155,9 @@ export const messages = {
                 update: 'Update',
             },
         },
-        stakingBanner: {
-            title: 'Grow your {networkSymbol} with staking',
-            subtitle: 'Earn rewards by staking your {networkSymbol} to support the network.',
+        stakingPromo: {
+            title: 'Grow your crypto with staking',
+            subtitle: 'Earn rewards by staking your crypto to support the network.',
             bannerText: 'Earn <bold>~{apy}% APY</bold> on your {networkSymbol}',
             bottomSheetTitle: 'Select account to stake',
         },
