+++ suite-native/module-home/package.json
@@ -16,6 +16,7 @@
         "@reduxjs/toolkit": "2.11.0",
         "@suite-common/graph": "workspace:*",
         "@suite-common/message-system": "workspace:*",
+        "@suite-common/redux-utils": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-core": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
@@ -29,6 +30,7 @@
         "@suite-native/device": "workspace:*",
         "@suite-native/device-manager": "workspace:*",
         "@suite-native/discovery": "workspace:*",
+        "@suite-native/feature-flags": "workspace:*",
         "@suite-native/firmware": "workspace:*",
         "@suite-native/graph": "workspace:*",
         "@suite-native/icons": "workspace:*",
@@ -36,6 +38,7 @@
         "@suite-native/link": "workspace:*",
         "@suite-native/navigation": "workspace:*",
         "@suite-native/settings": "workspace:*",
+        "@suite-native/staking": "workspace:*",
         "@suite-native/theme": "workspace:*",
         "@suite-native/toasts": "workspace:*",
         "@trezor/device-utils": "workspace:*",
