+++ suite-native/module-home/tsconfig.json
@@ -7,6 +7,9 @@
             "path": "../../suite-common/message-system"
         },
         {
+            "path": "../../suite-common/redux-utils"
+        },
+        {
             "path": "../../suite-common/wallet-config"
         },
         {
@@ -25,6 +28,7 @@
         { "path": "../device" },
         { "path": "../device-manager" },
         { "path": "../discovery" },
+        { "path": "../feature-flags" },
         { "path": "../firmware" },
         { "path": "../graph" },
         { "path": "../icons" },
@@ -32,6 +36,7 @@
         { "path": "../link" },
         { "path": "../navigation" },
         { "path": "../settings" },
+        { "path": "../staking" },
         { "path": "../theme" },
         { "path": "../toasts" },
         { "path": "../../packages/device-utils" },
